+++ test/groovy/UnityPackageOptionsTest.groovy
@@ -72,6 +72,7 @@ class UnityPackageOptionsTest {
         assertEquals('node:slim', defaults.prepareImage)
         assertEquals('linux && docker', defaults.publishAgent)
         assertEquals('node:slim', defaults.publishImage)
+        assertEquals('--network verdaccio', defaults.publishArgs)
         assertEquals([Unity: 'compose-unity'], defaults.unityAgents)
         assertEquals(['Library'], defaults.packageOptions.formattingExclusions)